refactor: move InfluxDB functionality and logging utilities into separate header files
