+++ include/utils.h
@@ -1,0 +1,10 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <Arduino.h>
+
+extern String logBuffer;
+
+void serialLog(String message);
+
+#endif // UTILS_H
