+++ src/utils.cpp
@@ -1,0 +1,9 @@
+#include "utils.h"
+
+String logBuffer = "";
+
+void serialLog(String message)
+{
+    logBuffer += message + "\n";
+    Serial.println(message);
+}
